+++ CHANGELOG.md
@@ -21,6 +21,7 @@ https://keepachangelog.com/ja/1.0.0/ に基づいて記述していく
 Unreleased
 -----------
 ### Added
+- add Frank matrix, tridiagonal Toeplitz matrix, 1D Laplacian matrix as sample matrices https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/199
 - add blas::copy https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/201
 - add fill function https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/193
 - add util::build_with functions https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/192